+++ src/main/java/cart/controller/dto/ProductRequest.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.dto;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
@@ -12,7 +12,8 @@ public class ProductRequest {
     @NotBlank
     private String imageUrl;
 
-    public ProductRequest() {}
+    public ProductRequest() {
+    }
 
     public ProductRequest(String name, Integer price, String imageUrl) {
         this.name = name;
